add zod demo and project

+++ src/routes/project.ts
@@ -1,10 +1,11 @@
-import { BaseProjectModel, Project, ProjectMetaData} from './../db_schema/project';
-import { auth, authAny, AuthenticatedRequest } from '../middleware/auth';
 import express from 'express';
+import { auth, authAny, AuthenticatedRequest } from '../middleware/auth';
+import { BaseProjectModel, Project, ProjectMetaData } from './../db_schema/project';
 
-import { ProjectModel } from '../db_schema/project';
 import { DemoModel } from '../db_schema/demo';
-import { defaultGeneralSetting } from '../db_schema/settings';
+import { IterationStepModel } from '../db_schema/iteration_step';
+import { PlanPropertyModel } from '../db_schema/plan-properties/plan_property';
+import { ProjectModel } from '../db_schema/project';
 
 export const projectRouter = express.Router();
 
@@ -20,7 +21,7 @@ projectRouter.post('/', auth, async (req: AuthenticatedRequest, res) => {
         }
 
         projectData.user = req.user._id;
-        delete projectData._id;
+        // delete projectData._id;
 
         const projectModel = new ProjectModel(projectData);
 
@@ -35,11 +36,7 @@ projectRouter.post('/', auth, async (req: AuthenticatedRequest, res) => {
         }
         projectId = newProject._id;
         
-        res.send({
-            status: true,
-            message: 'Project created',
-            data: newProject
-        });
+        res.send(newProject);
 
     } catch (ex : any) {
         console.log(ex.message);
@@ -69,11 +66,7 @@ projectRouter.put('/:id', auth, async (req, res) => {
 
         await project.save();
 
-        res.send({
-            status: true,
-            message: 'Project is stored',
-            data: project
-        });
+        res.send(project);
 
     } catch (ex : any) {
         res.send(ex.message);
@@ -90,9 +83,7 @@ projectRouter.get('', auth, async (req: AuthenticatedRequest, res) => {
         return res.status(404).send({ message: 'No project found.' });
     }
 
-    res.send({
-        data: projects
-    });
+    res.send(projects);
 
 });
 
@@ -105,17 +96,16 @@ projectRouter.get('/meta-data', auth, async (req: any, res) => {
         return res.status(404).send({ message: 'No project found.' });
     }
 
-    let metaDataList: ProjectMetaData[] = projects.map(project => ({
-        _id: project._id,
-        updated: project.updated,
-        name: project.name,
-        user: project.user,
-        description: project.description
-    }))
+    let metaDataList: ProjectMetaData[] = projects.filter(p => p._id !== undefined).
+        map(project => ({
+                _id: project._id,
+                public: project.public,
+                name: project.name,
+                user: project.user.toString(),
+            })
+        )   
 
-    res.send({
-        data: metaDataList
-    });
+    res.send(metaDataList);
 
 });
 
@@ -138,9 +128,7 @@ projectRouter.get('/:id', authAny, async (req: AuthenticatedRequest, res) => {
         if (project) { 
 
             if(req.user.role != 'user-study'){
-                return res.send({
-                    data: project
-                });
+                return res.send(project);
             }
             else{
                 return res.status(401).send()
@@ -151,9 +139,7 @@ projectRouter.get('/:id', authAny, async (req: AuthenticatedRequest, res) => {
         const demo = await DemoModel.findOne({ _id: id });
         if (demo) { 
 
-            return res.send({
-                data: demo
-            });
+            return res.send(demo);
         }
 
         return res.status(500).send({ message: 'No project found.' });
@@ -173,16 +159,16 @@ projectRouter.delete('/meta-data/:id', auth, async (req: AuthenticatedRequest, r
         const id = req.params.id;
 
         // // delete iteration steps
-        // const iterationsDeleteResult = await IterationStepModel.deleteMany({ project: id});
-        // if (!iterationsDeleteResult) { 
-        //     return res.status(404).send({ message: 'Problem during project deletion occurred' }); 
-        // }
+        const iterationsDeleteResult = await IterationStepModel.deleteMany({ project: id});
+        if (!iterationsDeleteResult) { 
+            return res.status(404).send({ message: 'Problem during project deletion occurred' }); 
+        }
 
         // // delete properties
-        // const propertyDeleteResult = await PlanPropertyModel.deleteMany({ project: id});
-        // if (!propertyDeleteResult) { 
-        //     return res.status(404).send({ message: 'Problem during project deletion occurred' }); 
-        // }
+        const propertyDeleteResult = await PlanPropertyModel.deleteMany({ project: id});
+        if (!propertyDeleteResult) { 
+            return res.status(404).send({ message: 'Problem during project deletion occurred' }); 
+        }
 
         // delete project itself
         const projectDeleteResult = await ProjectModel.deleteOne({ _id: id, user: req.user._id });
@@ -190,9 +176,8 @@ projectRouter.delete('/meta-data/:id', auth, async (req: AuthenticatedRequest, r
             return res.status(404).send({ message: 'No project found.' }); 
         }
 
-        res.send({
-            data: projectDeleteResult
-        });
+        res.send(true);
+
     } catch (ex : any) {
         console.log(ex);
         res.status(500).send();
